Minimize asm 0-padding in opcodes and mbc

+++ pyboy/core/mb.pxd
@@ -59,7 +59,7 @@ cdef class Motherboard:
     @cython.locals(pc=cython.int, bank=cython.int)
     cdef bint breakpoint_reached(self) noexcept
 
-    cdef uint8_t getitem(self, uint16_t) noexcept
+    cdef uint8_t getitem(self, uint64_t) noexcept
     cdef void setitem(self, uint16_t, uint8_t) noexcept
 
     @cython.locals(offset=cython.int, dst=cython.int, n=cython.int)
